refactor toplevel for ghci
